requirements.txt
fastapi==0.110.0
uvicorn==0.34.3
google-generativeai==0.4.1
requests==2.31.0
pydantic==2.11.7
numpy==1.26.4
pillow==10.2.0

